feat(custom-chart-web): refactor, sizing calculations
